+++ src/app/utils/dateFormat.ts
@@ -19,7 +19,7 @@ export const formatDate = (date: Date) => {
 };
 
 /**
- * 시간 포맷팅
+ * 시간 포맷팅 1
  * @param date 시간
  * @returns 시간 포맷팅
  */
@@ -29,6 +29,12 @@ export const formatTime = (date: Date) => {
   return `${hours}:${minutes}`;
 };
 
+/**
+ * 시간 포맷팅 2
+ * @param startTime 시간
+ * @param endTime 시간
+ * @returns 시간 포맷팅 00:00 - 00:00
+ */
 export const formatScheduleTimeOnly = (
   startTime: string,
   endTime: string
@@ -60,7 +66,6 @@ const getKoreanDay = (date: Date) => {
  * @param endTime 종료 시간
  * @returns 예: 2025년 8월 9일 (금) 17:30 - 18:30
  */
-
 export const formatSchedule = (startTime: string, endTime: string): string => {
   const startDate = new Date(startTime);
   const endDate = new Date(endTime);
@@ -78,6 +83,11 @@ export const formatSchedule = (startTime: string, endTime: string): string => {
   return `${year}년 ${month}월 ${day}일 (${dayOfWeek}) ${startHour}:${startMinute} - ${endHour}:${endMinute}`;
 };
 
+/**
+ * D-{DAY} 문자열로 변환
+ * @param startTime 시작 시간
+ * @returns 예: D-3, D+2, D-DAY
+ */
 export const getDDay = (startTime: string): string => {
   const today = new Date();
   const eventDate = new Date(startTime);
